// frontend/src/api/generated/model/foodDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * healthManager API
 * OpenAPI spec version: 1.0
 */
import type { FoodSource } from './foodSource';
import type { NutrientsDto } from './nutrientsDto';
import type { FoodUnitDto } from './foodUnitDto';

export interface FoodDto {
  id: string;
  name: string;
  /** @nullable */
  brand: string | null;
  source: FoodSource;
  /** Per 100 g */
  per100g: NutrientsDto;
  units: FoodUnitDto[];
}